Move unwrapBean() to ScriptSourceAwareImplementationIdentifier
