+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/properties/ScriptSourceAwareImplementationIdentifier.java
@@ -16,16 +16,47 @@
 
 package org.gradle.api.internal.tasks.properties;
 
+import com.google.common.annotations.VisibleForTesting;
+import org.codehaus.groovy.runtime.ConvertedClosure;
 import org.gradle.internal.properties.bean.ImplementationIdentifier;
 import org.gradle.internal.scripts.ScriptOriginUtil;
 import org.gradle.internal.snapshot.impl.ImplementationValue;
+import org.gradle.util.internal.ClosureBackedAction;
 import org.gradle.util.internal.ConfigureUtil;
+
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.Proxy;
 
 public class ScriptSourceAwareImplementationIdentifier implements ImplementationIdentifier {
     @Override
     public ImplementationValue identify(Object bean) {
-        Object unwrapped = ConfigureUtil.unwrapBean(bean);
+        Object unwrapped = unwrapBean(bean);
         String classIdentifier = ScriptOriginUtil.getOriginClassIdentifier(unwrapped);
         return new ImplementationValue(classIdentifier, unwrapped);
+    }
+
+    @VisibleForTesting
+    static Object unwrapBean(Object bean) {
+        // When Groovy coerces a Closure into an SAM type, then it creates a Proxy which is backed by the Closure.
+        // We want to track the implementation of the Closure, since the class name and classloader of the proxy will not change.
+        // Java and Kotlin Lambdas are coerced to SAM types at compile time, so no unpacking is necessary there.
+        if (Proxy.isProxyClass(bean.getClass())) {
+            InvocationHandler invocationHandler = Proxy.getInvocationHandler(bean);
+            if (invocationHandler instanceof ConvertedClosure) {
+                return ((ConvertedClosure) invocationHandler).getDelegate();
+            }
+            return invocationHandler;
+        }
+
+        // Same as above, if we have wrapped a closure in a WrappedConfigureAction or a ClosureBackedAction, we want to
+        // track the closure itself, not the action class.
+        if (bean instanceof ConfigureUtil.WrappedConfigureAction) {
+            return ((ConfigureUtil.WrappedConfigureAction<?>) bean).getConfigureClosure();
+        }
+
+        if (bean instanceof ClosureBackedAction) {
+            return ((ClosureBackedAction<?>) bean).getClosure();
+        }
+        return bean;
     }
 }
